Fix type annotation

+++ regression/python/function-return-if/main.py
@@ -1,8 +1,8 @@
 def func() -> int:
    return 1
 
-x : int = func()
+x:int = func()
 if (x == 0):
-    y : int = 1/x
+    y:float = 1/x
 else:
     assert (x == 0)
